msm: mdss: hdmi: add checksum for HDR infoframe
Checksum for the HDR infoframe is set to zero by default
as this is not a mandatory field as per the HDMI spec.

However certain HDMI sinks still expect a non-zero
checksum. Otherwise they disregard the infoframe
and the sink does not enter HDR mode despite other fields
of the infoframe being valid.

Add a valid checksum to the HDR infoframe to improve
interoperability of our HDR solution for HDMI.

Change-Id: Ic715981c458384e75f3ab1af69080478ab852d37
Signed-off-by: Narender Ankam <nankam@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>